requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
djangorestframework==3.3.1
gevent==1.1b5
gevent-socketio==0.3.6
gevent-websocket==0.9.5
google-api-python-client==1.4.2
greenlet==0.4.9
gunicorn==19.3.0
httplib2==0.9.2
mysqlclient==1.3.6
oauth2==1.9.0.post1
oauth2client==1.5.1
oauthlib==1.0.3
Pillow==3.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
Pygments==2.0.2
PyJWT==1.4.0
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.13
python3-openid==3.0.7
redis==2.10.3
requests==2.8.0
requests-oauthlib==0.5.0
rsa==3.2
selenium==2.49.1
simplejson==3.8.0
six==1.10.0
sockjs-tornado==1.0.2
static3==0.6.1
SwampDragon==0.4.2.2
tornado==4.2.1
tornado-redis==2.4.18
uritemplate==0.6
virtualenv==13.1.2
websockets==2.6
whitenoise==2.0.6
